feat: Comment 단건 조회 추가 및 좋아요 개수 필드 추가

+++ src/main/java/com/sparta/sixhundredbills/comment_like/dto/CommentLikeResponseDto.java
@@ -12,5 +12,5 @@ public class CommentLikeResponseDto {
     private String message; // 응답 메시지
     private Long commentId; // 댓글 ID
     private Long userId; // 사용자 ID
-    private int likesCount; // 좋아요 수
+    private Long likesCount; // 좋아요 수
 }
